Move scalar kernels out of test/
* CMakeLists.txt: Add cvt and gemm kernels. Remove unused source list.
* include/skl-common.h: Add back util function without memcpy
  dependency.
* include/skl-ref.h: Add new headers. Remove Zfh guard.
* ref/cvt/cvt_f4e2m1_f8e4m3_scalar.c,
* ref/cvt/cvt_f4e2m1_f8e4m3_scalar.h,
* ref/cvt/cvt_ofp8_scalar.c
* ref/cvt/cvt_ofp8_scalar.h
* ref/gemm/gemm_f8e4m3rc_f8e4m3rc_f32rc_scalar.c,
* ref/gemm/gemm_f8e4m3rc_f8e4m3rc_f32rc_scalar.h,
* ref/gemm/gemm_f8e4m3rcprc_f8e4m3rcprc_f32rcprc_scalar.c,
* ref/gemm/gemm_f8e4m3rcprc_f8e4m3rcprc_f32rcprc_scalar.h: New files.
* test/cvt/cvt_ofp_zvfofp4min.c,
* test/cvt/cvt_ofp_zvfofp8min.c,
* test/cvt/cvt_ofp_zvfofp8min_zvfbfmin.c: Include skl-ref.h.
* test/gemm/gemm_f8e4m3rc_f8e4m3rc_f32rc.c,
* test/gemm/gemm_f8e4m3rcprc_f8e4m3rcprc_f32rcprc.c: Include skl-ref.h.
  Remove scalar kernels.
* test/skl-test.h: Remove scalar kernels.
