+++ include/skl-ref.h
@@ -10,6 +10,7 @@
  */
 
 #include "../ref/gemm/gemm_bf16rc_bf16rc_f32rc_scalar.h"
+#include "../ref/gemm/gemm_f8e4m3rc_f8e4m3rc_f32rc_scalar.h"
 #include "../ref/gemm/gemm_f16rc_f16rc_f16rc_scalar.h"
 #include "../ref/gemm/gemm_f16rc_f16rc_f32rc_scalar.h"
 #include "../ref/gemm/gemm_f32rc_f32rc_f32rc_scalar.h"
@@ -17,6 +18,7 @@
 #include "../ref/gemm/gemm_i8rc_i8rc_i32rc_scalar.h"
 
 #include "../ref/gemm/gemm_bf16rcprc_bf16rcprc_f32rcprc_scalar.h"
+#include "../ref/gemm/gemm_f8e4m3rcprc_f8e4m3rcprc_f32rcprc_scalar.h"
 #include "../ref/gemm/gemm_f16rcprc_f16rcprc_f32rcprc_scalar.h"
 #include "../ref/gemm/gemm_f32rcprc_f32rcprc_f32rcprc_scalar.h"
 #include "../ref/gemm/gemm_f64rcprc_f64rcprc_f64rcprc_scalar.h"
@@ -38,12 +40,16 @@
 #include "../ref/logistic/logistic_f32_scalar.h"
 
 /*
+ * Conversion Functions
+ */
+#include "../ref/cvt/cvt_f4e2m1_f8e4m3_scalar.h"
+#include "../ref/cvt/cvt_ofp8_scalar.h"
+
+/*
  * Depthwise Convolution Kernels
  */
 
-#if defined(__riscv_zfh)
 #include "../ref/depthwise_conv2d/depthwise_conv2d_f16_f16_f16_scalar.h"
-#endif
 
 #include "../ref/depthwise_conv2d/depthwise_conv2d_f32_f32_f32_scalar.h"
 #include "../ref/depthwise_conv2d/depthwise_conv2d_i8_i8_i32_scalar.h"
