+++ test/cvt/cvt_ofp_zvfofp8min_zvfbfmin.c
@@ -1,4 +1,4 @@
-// Copyright 2025 SiFive, Inc.
+// Copyright 2026 SiFive, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
 #if !defined(__riscv_zvfofp8min) || !defined(__riscv_zvfbfmin)
@@ -25,6 +25,8 @@
 #define SATURATION_TEST_MODE 3 // 1: no saturation, 2: saturation, 3: both
 #endif
 
+// NOLINTNEXTLINE(misc-include-cleaner)
+#include "skl-ref.h"
 #include "skl-test.h"
 #include "skl.h"
